refactor: sidebar description partials moved outside folder, aligned imports
