+++ src/components/v5/common/ActionSidebar/partials/ActionSidebarDescription/partials/CreateNewDomainDescription.tsx
@@ -4,8 +4,9 @@ import { useFormContext } from 'react-hook-form';
 import { ColonyActionType } from '~gql';
 import { formatText } from '~utils/intl';
 
-import { CreateNewTeamFormValues } from '../../../forms/CreateNewTeamForm/consts';
-import CurrentUser from '../CurrentUser/CurrentUser';
+import { CreateNewTeamFormValues } from '../../forms/CreateNewTeamForm/consts';
+
+import CurrentUser from './CurrentUser';
 
 const displayName =
   'v5.common.ActionsSidebar.partials.ActionSidebarDescription.partials.CreateNewDomainDescription';
