+++ src/components/v5/common/ActionSidebar/partials/ActionSidebarDescription/partials/ManageColonyObjectiveDescription.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { ExtendedColonyActionType } from '~types';
 import { formatText } from '~utils/intl';
 
-import CurrentUser from '../CurrentUser/CurrentUser';
+import CurrentUser from './CurrentUser';
 
 const displayName =
   'v5.common.ActionsSidebar.partials.ActionSidebarDescription.partials.ManageColonyObjectiveDescription';
