+++ src/components/v5/common/ActionSidebar/partials/ActionSidebarDescription/partials/SimplePaymentDescription.tsx
@@ -8,39 +8,38 @@ import Numeral from '~shared/Numeral';
 import { formatText } from '~utils/intl';
 import { getTokenDecimalsWithFallback } from '~utils/tokens';
 
-import { TransferFundsFormValues } from '../../../forms/TransferFundsForm/hooks';
-import CurrentUser from '../CurrentUser/CurrentUser';
+import { SimplePaymentFormValues } from '../../forms/SimplePaymentForm/hooks';
+
+import CurrentUser from './CurrentUser';
+import RecipientUser from './RecipientUser';
 
 const displayName =
-  'v5.common.ActionsSidebar.partials.ActionSidebarDescription.partials.TransferFundsDescription';
+  'v5.common.ActionsSidebar.partials.ActionSidebarDescription.partials.SimplePaymentDescription';
 
-export const TransferFundsDescription = () => {
-  const {
-    colony: { domains, nativeToken, tokens },
-  } = useColonyContext();
-  const formValues = useFormContext<TransferFundsFormValues>().getValues();
+export const SimplePaymentDescription = () => {
+  const formValues = useFormContext<SimplePaymentFormValues>().getValues();
+  const { colony } = useColonyContext();
+  const { amount: { amount, tokenAddress } = {}, recipient } = formValues;
 
-  const { amount: { amount, tokenAddress } = {}, from, to } = formValues;
-
-  const fromDomain = domains?.items.find((domain) => domain?.nativeId === from);
-  const toDomain = domains?.items.find((domain) => domain?.nativeId === to);
-  const matchingColonyToken = tokens?.items.find(
+  const { nativeToken } = colony;
+  const matchingColonyToken = colony.tokens?.items.find(
     (colonyToken) => colonyToken?.token?.tokenAddress === tokenAddress,
   );
+
+  const recipientUser = <RecipientUser userAddress={recipient} />;
 
   return (
     <>
       {formatText(
         { id: 'action.title' },
         {
-          actionType: ColonyActionType.MoveFunds,
-          fromDomain: fromDomain?.metadata ? fromDomain.metadata.name : '',
-          toDomain: toDomain?.metadata ? toDomain.metadata.name : '',
+          actionType: ColonyActionType.Payment,
           tokenSymbol: matchingColonyToken
             ? matchingColonyToken.token.symbol
             : formatText({
                 id: 'actionSidebar.metadataDescription.tokens',
               }),
+          recipient: recipientUser,
           amount: amount ? (
             <Numeral
               value={moveDecimal(
@@ -51,7 +50,7 @@ export const TransferFundsDescription = () => {
             />
           ) : (
             formatText({
-              id: 'actionSidebar.metadataDescription.funds',
+              id: 'actionSidebar.metadataDescription.anAmount',
             })
           ),
           initiator: <CurrentUser />,
@@ -61,5 +60,5 @@ export const TransferFundsDescription = () => {
   );
 };
 
-TransferFundsDescription.displayName = displayName;
-export default TransferFundsDescription;
+SimplePaymentDescription.displayName = displayName;
+export default SimplePaymentDescription;
